fix: Service의 특성에 따른 Read&Write 분리
